bugfix files check and removal

+++ inc/model/cli/files.php
@@ -33,6 +33,12 @@ final class files extends \fpcm\model\abstracts\cli {
     private array $filesIndex = [];
 
     /**
+     * Directory index array
+     * @var array
+     */
+    private array $dirsIndex = [];
+
+    /**
      * Folder excludes parameter
      * @var array
      */
@@ -196,7 +202,6 @@ final class files extends \fpcm\model\abstracts\cli {
 
         foreach ($this->filesIndex as $path) {
 
-
             $progress
                     ->setOutputText(\fpcm\model\files\ops::removeBaseDir($path))
                     ->setCurrentValue($i)
@@ -207,6 +212,11 @@ final class files extends \fpcm\model\abstracts\cli {
                 continue;
             }
 
+            if (is_dir($path)) {
+                $this->dirsIndex[] = $path;
+                continue;
+            }
+
             if (!unlink($path)) {
                 $this->output(sprintf('Failed to remove files %s, cancel process...', $path), true);
             }
@@ -214,8 +224,29 @@ final class files extends \fpcm\model\abstracts\cli {
             $i++;
         }
         
+        unset($path);
+        
+        if (count($this->dirsIndex)) {
+            
+            foreach ($this->dirsIndex as $path) {
+                
+                $progress
+                        ->setOutputText(\fpcm\model\files\ops::removeBaseDir($path))
+                        ->setCurrentValue($i)
+                        ->output();                
+
+                if (!\fpcm\model\files\ops::deleteRecursive($path)) {
+                    $this->output(sprintf('Failed to remove folder %s, cancel process...', $path), true);
+                }
+
+                $i++;
+            }
+
+        }
+        
         usleep(5000);
         $this->filesIndex = [];
+        $this->dirsIndex = [];
 
         $this->check();
     }
